Replace Birds simulation with individual Bird objects
This is a preparatory step towards a componentised sim with simulation
objects each being composed of their own sets of behaviour and drawing
routines.

This is so I can support things like obstacles and suchlike, and can put
them in the simulated world with different behaviour to thihngs like
birds.

At the moment the Bird class is still one honking great thing, but I can
break this out into guidance, physics and other components subsequently.