Extract goal object
Lets the world really be just a bag of state, and the objects in the
system be constructed by the main app.

Residually unhappy with the way this is still breaking
dimension-independence (the game objects are inherently 3D at the
moment) but I think this could be parameterised easily enough once I get
the graphics components of the game objects extracted.

Also not really sure about the way World has a bunch of names for
things, e.g. `birds` and `goal`, which the game objects (ok birds) then
use to react to things differently. We'll see I guess.